requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
trl>=0.7.0

# Unsloth for efficient fine-tuning
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git

# Evaluation metrics
rouge-score>=0.1.0
sacrebleu>=2.3.0
evaluate>=0.4.0
nltk>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
pyyaml>=6.0.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Memory optimization
bitsandbytes>=0.41.0
xformers>=0.0.22

# Additional dependencies for specific tasks
sentencepiece>=0.1.99
protobuf>=4.24.0
